upd - replase expressions from TodoFooter to App
