+++ src/components/TodoItem/TodoItem.tsx
@@ -22,7 +22,6 @@ export const TodoItem: FC<Props> = props => {
           type="checkbox"
           className="todo__status"
           checked={todo.completed}
-          onChange={() => {}}
         />
       </label>
